% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{as_check}
\alias{as_check}
\title{Create a checker}
\usage{
as_check(x, ...)
}
\arguments{
\item{x}{object to convert to check}

\item{...}{other arguments passed onto methods}
}
\description{
Create a checker
}
