% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{as_status}
\alias{as_status}
\title{Convert return value to status}
\usage{
as_status(x, ...)
}
\description{
This is not the same as \code{\link[vctrs:vec_cast]{vctrs::vec_cast()}} to \code{chex_status}. These are the
types that \code{do_check()} will recognize as valid return values.
}
\keyword{internal}
