% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_that.R
\name{check_that}
\alias{check_that}
\title{Run some checks}
\usage{
check_that(x, ...)
}
\arguments{
\item{x}{An object to run your checks against}

\item{...}{Checks to run against the object}
}
\description{
Run some checks
}
