% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chex-package.R
\docType{package}
\name{chex-package}
\alias{chex}
\alias{chex-package}
\title{chex: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Shun Sambongi \email{shun.sambongi@biofiredx.com}

}
\keyword{internal}
