% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chex-package.R, R/result.R
\name{chex-vctrs}
\alias{chex-vctrs}
\alias{vec_ptype2.chex_result}
\alias{vec_cast.chex_result}
\title{Internal vctrs functions}
\usage{
\method{vec_ptype2}{chex_result}(x, y, ...)

\method{vec_cast}{chex_result}(x, to, ...)
}
\description{
Internal vctrs functions
}
\keyword{internal}
