% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{data}
\alias{data}
\alias{status}
\alias{status<-}
\alias{description}
\alias{description<-}
\alias{reason}
\alias{reason<-}
\alias{duration}
\alias{duration<-}
\title{Get/set data fields}
\usage{
status(x, ...)

status(x) <- value

description(x, ...)

description(x) <- value

reason(x, ...)

reason(x) <- value

duration(x, ...)

duration(x) <- value
}
\arguments{
\item{x}{object used to get or set fields}

\item{...}{other args passed on to methods}

\item{value}{new value to set}
}
\description{
Get/set data fields
}
