% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/status.R
\docType{data}
\name{preset-status}
\alias{preset-status}
\alias{SKIP}
\alias{PASS}
\alias{OTHER}
\alias{WARNING}
\alias{FAIL}
\alias{ERROR}
\alias{INVALID}
\title{Preset Statuses}
\format{













}
\usage{
SKIP

PASS

OTHER

WARNING

FAIL

ERROR

INVALID
}
\description{
Preset Statuses
}
\keyword{datasets}
