% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{set_description}
\alias{set_description}
\title{Set object description}
\usage{
set_description(x, description)
}
\arguments{
\item{x}{object being described}

\item{description}{the description}
}
\description{
Set object description
}
