% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_that.R
\name{skip_check}
\alias{skip_check}
\title{Skip the current check}
\usage{
skip_check(details = NULL)
}
\arguments{
\item{details}{A character vector of size 1 with any details about why the
check was skipped}
}
\description{
Call this function from inside of a check function to stop checking and quit
the check early.
}
