% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/status.R
\name{status}
\alias{status}
\alias{is_status}
\alias{as_status}
\title{\code{status} vector}
\usage{
status(x)

is_status(x)

as_status(x)
}
\arguments{
\item{x}{\itemize{
\item For \code{status()}: A character vector
\item For \code{is_status()}: An object to test
\item For \code{as_status()}: An object to cast
}}
}
\description{
This creates a character vector that represents the result status of a check.
Any status that is created will be changed to all uppercase, i.e.
\code{status("pass") == status("PASS")}.
}
