% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chex-package.R
\docType{package}
\name{chex-package}
\alias{chex}
\alias{chex-package}
\title{chex: Perform Checks}
\description{
Checks stuff.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/shunsambongi/chex}
  \item Report bugs at \url{https://github.com/shunsambongi/chex/issues}
}

}
\author{
\strong{Maintainer}: Shun Sambongi \email{shundev3@gmail.com}

}
\keyword{internal}
