% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result.R
\name{result}
\alias{result}
\alias{as_result}
\alias{is_result}
\title{\code{result} vector}
\usage{
result(
  status,
  details = chex::details(status) \%||\% NA_character_,
  description = NA_character_
)

as_result(x, ...)

is_result(x)
}
\arguments{
\item{status}{the status of the result}

\item{details}{any details explaining the status}

\item{description}{description of the check which produced the result}

\item{x}{\itemize{
\item For \code{as_result()}: an object to cast
\itemize{
\item For \code{is_result()}: an object to check
}
}}

\item{...}{other arguments passed onto methods}
}
\description{
\code{result} vector
}
