% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result.R
\name{result}
\alias{result}
\alias{as_result}
\alias{is_result}
\title{\code{result} vector}
\usage{
result(
  status,
  details = chex::details(status) \%||\% NA_character_,
  description = NA_character_
)

as_result(x, ...)

is_result(x)
}
\description{
\code{result} vector
}
